Adds code to draw a birthday cake using turtle graphics
This commit adds code to draw a birthday cake using turtle graphics. The code includes functions to draw the cake layers, candles, sticks, and toppings. It also includes functions to write "Happy" and "Birthday" inside a circle on top of the cake. The cake is drawn with different colors and sizes for each layer, and the toppings are drawn in various colors. This code can be used to create a visual representation of a birthday celebration.